Fix the requested things from PR #75
